refactor(cli): group the run commands into one directory
run.ts and runs.ts differed by a single character and sat next to each other,
which is worse than the run-agent.ts they replaced — renaming the file after its
command fixed consistency and broke legibility. Alongside them sat
one-shot-envelope.ts and one-shot-flags.ts, wearing a jargon prefix only to say
which command they belonged to.

The directory says that instead, so each file can be named for what it holds:

  commands/run/execute.ts     jazz run
  commands/run/lifecycle.ts   jazz runs
  commands/run/envelope.ts    the output contract
  commands/run/flags.ts       closed-set flag validation

Splits the one test file to match, and moves the envelope key-order assertion
out of the conversation-persistence block it had been filed under — it tests the
published shape, not persistence.

Drops the auth-required run state. Nothing detects a missing credential, so
nothing could ever enter it, and an unreachable state that still shows up in the
store and the CLI's formatter is the same dead vocabulary as the TaskNotFound
errors nothing throws. The lifecycle takes one union member and one transition
row to extend when there is a caller.
