+++ src/cli/commands/run/envelope.test.ts
@@ -1,0 +1,83 @@
+import { describe, expect, it } from "bun:test";
+import { formatOneShotError, formatOneShotResult, type OneShotSuccess } from "./envelope";
+
+const baseResult: OneShotSuccess = {
+  answer: "Hello from the agent",
+  costUSD: 0.0012,
+  tokenUsage: { promptTokens: 10, completionTokens: 5, totalTokens: 15 },
+  toolCalls: [{ id: "call_1", name: "web_search", arguments: '{"q":"x"}' }],
+};
+
+describe("formatOneShotResult", () => {
+  it("plain mode emits only the trimmed answer with a trailing newline", () => {
+    const output = formatOneShotResult({ ...baseResult, answer: "  Hello  \n\n" }, { json: false });
+    expect(output).toBe("Hello\n");
+  });
+
+  it("plain mode does not include header, footer, or JSON envelope keys", () => {
+    const output = formatOneShotResult(baseResult, { json: false });
+    expect(output).not.toContain("◉");
+    expect(output).not.toContain("completed");
+    expect(output).not.toContain('"ok"');
+  });
+
+  it("json mode emits exactly one single-line envelope", () => {
+    const output = formatOneShotResult(baseResult, { json: true });
+    expect(output.endsWith("\n")).toBe(true);
+    expect(output.trimEnd().includes("\n")).toBe(false);
+    expect(JSON.parse(output)).toEqual({
+      ok: true,
+      answer: "Hello from the agent",
+      costUSD: 0.0012,
+      tokenUsage: { promptTokens: 10, completionTokens: 5, totalTokens: 15 },
+      toolCalls: [{ id: "call_1", name: "web_search", arguments: '{"q":"x"}' }],
+    });
+  });
+
+  it("omits messages by default", () => {
+    const output = formatOneShotResult(baseResult, { json: true });
+    expect(JSON.parse(output)).not.toHaveProperty("messages");
+  });
+
+  it("includes messages when set (--ephemeral round-trip)", () => {
+    const messages: ChatMessage[] = [
+      { role: "user", content: "hi" },
+      { role: "assistant", content: "hello" },
+    ];
+    const output = formatOneShotResult({ ...baseResult, messages }, { json: true });
+    expect(JSON.parse(output).messages).toEqual(messages);
+  });
+});
+
+describe("the success envelope's key order", () => {
+  // Bridges outside this repo read these keys. Adding one is a compatible change; the
+  // shape changing under them without anyone noticing is not.
+  it("is exactly ok, answer, costUSD, tokenUsage, toolCalls", () => {
+    const output = formatOneShotResult(baseResult, { json: true });
+    expect(Object.keys(JSON.parse(output))).toEqual([
+      "ok",
+      "answer",
+      "costUSD",
+      "tokenUsage",
+      "toolCalls",
+    ]);
+  });
+});
+
+describe("formatOneShotError", () => {
+  it("plain mode emits the message with a trailing newline", () => {
+    expect(formatOneShotError("Agent not found", { json: false })).toBe("Agent not found\n");
+  });
+
+  it("json mode emits an ok:false envelope including costUSD", () => {
+    expect(JSON.parse(formatOneShotError("boom", { json: true }, 0.5))).toEqual({
+      ok: false,
+      error: "boom",
+      costUSD: 0.5,
+    });
+  });
+
+  it("json mode defaults costUSD to 0", () => {
+    expect(JSON.parse(formatOneShotError("boom", { json: true })).costUSD).toBe(0);
+  });
+});
